requirements.txt
fastapi
tortoise-orm
databases
aiosqlite
pydantic
pytest
pytest-asyncio
uvicorn
httpx
